Move dtos to separate folder
