+++ src/Application/Beers/Queries/GetBeer/GetBeerQuery.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Beers.Dtos;
+using MediatR;
 
 namespace Application.Beers.Queries.GetBeer;
